server/requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.5.3
six==1.11.0
SQLAlchemy==1.1.14
uWSGI==2.0.15
Werkzeug==0.12.2
